refactor(client): rename boxes to contentBoxComponents; move OverviewCardGrid to layout
The boxes/ folder mixed the content-box family (ContentBox + the entity
cards + DashboardCard) with OverviewCardGrid, a section-landing nav grid.
Rename the folder to contentBoxComponents/ to describe what it holds, and
move OverviewCardGrid into layout/ alongside the other page chrome.

To avoid churning the ~30 importers (notably the 12 dashboard files using
DashboardCard), boxes/ keeps a per-file re-export shim for each content-box
component, so existing @/components/boxes/* import paths still resolve.
OverviewCardGrid's 3 route importers are repointed directly since it leaves
boxes/ entirely.
