+++ packages/client/src/components/contentBoxComponents/ContentBox.stories.tsx
