+++ packages/client/src/components/contentBoxComponents/CourseBox.stories.tsx
